+++ diskstat/main.py
@@ -1,23 +1,30 @@
-from diskstat.disks import main_console
+from diskstat.gen import gen
+import diskstat.autostart as autostart
 import argparse
-from diskstat.program_window import App
-import sys
+
 
 def main():
+    gen()
     parser = argparse.ArgumentParser()
 
     parser.add_argument(
-        "-c", "--console", action="store_true", default=False, help="console mode"
+        "-d", "--disable", action="store_true", default=False, help="disable auto start"
+    )
+    parser.add_argument(
+        "-o", "--open", action="store_true", default=False, help="open startup folder"
     )
 
     args = parser.parse_args()
+    autostart.enable()
 
-    if args.console:
-        main_console()
-    else:
-        app = App(sys.argv)
-        app.window.show()
-        app.exec()
+    if args.disable:
+        autostart.disable()
+
+    if args.open:
+        autostart.open_start_folder()
+
+    if not args.disable:
+        autostart.exec()
 
 
 if __name__ == "__main__":
